+++ tests/src/lib.rs
@@ -18,6 +18,7 @@ use tarantool::ffi::lua as ffi_lua;
 use tarantool::index::IndexType;
 use tarantool::space::{Field, FieldType, Space};
 
+mod auth;
 mod r#box;
 mod coio;
 mod common;